pyproject.toml
[project]
name = "reddit-researcher"
version = "0.1.0"
description = "CLI tools for researching Reddit data"
authors = [{name = "Derious Vaughn"}]
requires-python = ">=3.12"
readme = "README.md"
dependencies = ["praw", "python-dotenv", "tqdm", "pyyaml"]

[project.scripts]
reddit-researcher = "reddit_researcher.cli.cli:main"
reddit-probe = "reddit_researcher.core.probe:main"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = 3.12
strict = true
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
